MISRA

+++ HAL/LCD/LCD.c
@@ -74,15 +74,11 @@ void LCD_DispChar(uint8 data) {
     // high-to-low pulse (450 ns wide) to E pin
     // wait 100 us
 
-
-
-    // MISRA :note 9226: integral expression of underlying type 'short' cannot be implicitly converted to type 'uint8' (aka 'unsigned char') because it is a complex expression [MISRA 2004 Rule 10.1, required]
-    // MISRA :note 9233: bitwise operator >> may not be applied to operand with signed underlying type [MISRA 2004 Rule 12.7, required]
-    uint8 HighNibb = (data & 0xf0) >> 4;
-    uint8 D01 = (HighNibb & 0x01);
-    uint8 D02 = ((HighNibb >> 1) & 0x01);
-    uint8 D03 = ((HighNibb >> 2) & 0x01);
-    uint8 D04 = ((HighNibb >> 3) & 0x01);
+    uint8 HighNibb = (data & (uint8)0xf0) >> 4;
+    uint8 D01 = (HighNibb & (uint8)0x01);
+    uint8 D02 = ((HighNibb >> (uint8)1) & (uint8)0x01);
+    uint8 D03 = ((HighNibb >> (uint8)2) & (uint8)0x01);
+    uint8 D04 = ((HighNibb >> (uint8)3) & (uint8)0x01);
     DIO_write(LCD_PORT, LCD_D4, D01);
     DIO_write(LCD_PORT, LCD_D5, D02);
     DIO_write(LCD_PORT, LCD_D6, D03);
@@ -91,15 +87,11 @@ void LCD_DispChar(uint8 data) {
     DIO_write(LCD_PORT, LCD_RS, 1);  //enable LCD to receive data
     send_falling_edge();             //send falling edge
 
-
-
-    // MISRA :note 9233: bitwise operator & may not be applied to operand with signed underlying type [MISRA 2004 Rule 12.7, required]
-    // MISRA :note 9225: integral expression of underlying type 'signed char' cannot be implicitly converted to type 'uint8' (aka 'unsigned char') because it is not a wider integer type of the same signedness [MISRA 2004 Rule 10.1, required]
-    uint8 LowNibb = data & 0x0f;
-    D01 = (LowNibb & 0x01);
-    D02 = ((LowNibb >> 1) & 0x01);
-    D03 = ((LowNibb >> 2) & 0x01);
-    D04 = ((LowNibb >> 3) & 0x01);
+    uint8 LowNibb = data & (uint8)0x0f;
+    D01 = (LowNibb & (uint8)0x01);
+    D02 = ((LowNibb >> (uint8)1) & (uint8)0x01);
+    D03 = ((LowNibb >> (uint8)2) & (uint8)0x01);
+    D04 = ((LowNibb >> (uint8)3) & (uint8)0x01);
     DIO_write(LCD_PORT, LCD_D4, D01);
     DIO_write(LCD_PORT, LCD_D5, D02);
     DIO_write(LCD_PORT, LCD_D6, D03);
@@ -112,16 +104,15 @@ void LCD_Print(uint8* str) {
     {
         LCD_DispChar(*str);  // print the characters of the string
 
-        // MISRA :note 9017: incrementing pointer [MISRA 2004 Rule 17.4, required]
-        str++;               // make the pointer points to the next character
+        // MISRA : we have a check so it doesn't go out of index
+        str++;               
     }
 }
 
 void LCD_PrintString(uint8* str, uint8 len) {
     for (uint8 itr = 0; itr < len;itr++) //keep in the loop until the end of the string
     {
-
-        // MISRA :note 9264: array subscript applied to variable 'str' declared with non-array type 'uint8 *' (aka 'unsigned char *') [MISRA 2004 Rule 17.4, required]
+        // MISRA : we have a check so it doesn't go out of index
         LCD_DispChar(str[itr]);  // print the characters of the string
     }
 }
